Explicitly set encoding on responses from CTS API
This one was tricky to track down. The German translation of Divān
was being incorrected detected as CP949. This led to mojibake when
not expected.

By explicitly setting the *default* encoding (by first checking) we
reduce a call to chardet which is quite expensive in my experience.

Nautilus should send back explicit encoding, but that is a different
issue for a different day.