Replace quadratic algorithm in cts.GetValidReff.
When the _debug parameter is set to True, as it is when called
by hooktest, the GetValidReff method for CapitainsCtsText files
used a quadratic algorithm to search for duplicate passage
reference strings. When a large number of passage references
is generated for a document this caused a significant processing
delay.

Instead, use a more verbose but O(N) algorithm.