Corrected a problem that subtype was sometimes not appearing on some objects
The problem was that if the first time a collection was processed it did not have a dc:type in its XML, this would not be saved on the object. Then, when it was parsed with all of its metadata, the subtype would not be added. This may also have been happening with other metadata, but I only noticed it on the subtype.